Portfolio Project 3 Case Study
Participated in Google UX Design certificate course. This was the last project done in that course. The goal was to design first an app and then a responsive website. I was able to choose the subject from three topics and I picked this one. This case study and its content is made according instructions given in the course.